Add comments

+++ src/Data_parser/Query_for_get_data.java
@@ -3,8 +3,8 @@ public class Query_for_get_data {
     
     private final String address = "https://api.hh.ru/vacancies?text=";
     
-    public String getQuery(int area, int specialization){
-        return String.format(address + "&specialization=%d&area=%d&per_page=100&page=", area, specialization);
+    public String getQuery(double area, double specialization){
+        return String.format(address + "&specialization=%f&area=%f&per_page=100&page=", area, specialization);
     }
     
 }
